% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_fu.R
\name{brazil_fu}
\alias{brazil_fu}
\title{Get Brazilian federative unit abbreviations}
\usage{
brazil_fu(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names or numeric codes of Brazilian regions or federal units.
Municipality codes are also supported. If \code{NULL},
returns a \code{\link[base:character]{character}} vector with all Brazilian
federal unit abbreviations (default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the abbreviations of
Brazilian federal units.
}
\description{
\code{brazil_fu()} returns a vector with the abbreviations of
Brazilian federal units.
}
\details{
Data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}, n.d.).
}
\examples{
brazil_fu()

brazil_fu("sp")
#> [1] "SP" # Expected

brazil_fu("sao paulo")
#> [1] "SP" # Expected

brazil_fu(35)
#> [1] "SP" # Expected

brazil_fu(3550308) # Municipality of São Paulo
#> [1] "SP" # Expected

brazil_fu(35503081) # >7 digits
#> [1] NA # Expected

brazil_fu(39027001) # Non-existent state code
#> [1] NA # Expected

brazil_fu("southeast")
#> [1] "ES" "MG" "RJ" "SP" # Expected

brazil_fu(3)
#> [1] "ES" "MG" "RJ" "SP" # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_municipality}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region}()},
\code{\link{brazil_region_code}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_code}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
