% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_municipality.R
\name{brazil_municipality}
\alias{brazil_municipality}
\title{Get Brazilian municipalities data}
\usage{
brazil_municipality(
  municipality = NULL,
  state = NULL,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  coords_method = "geobr",
  force = FALSE
)
}
\arguments{
\item{municipality}{(optional) A \code{\link[base:character]{character}} vector
with the name of the municipalities. If \code{NULL} the function returns all
municipalities (default: \code{NULL}).}

\item{state}{(optional) A \code{\link[base:character]{character}} vector with the
name of the states (default: \code{NULL}).}

\item{year}{(optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(default: \code{Sys.Date() |> substr(1, 4) |> as.numeric()}).}

\item{coords_method}{(optional) A string indicating the method to retrieve
the latitude and longitude coordinates of the municipalities
(default: \code{"geobr"}). Options are:
\itemize{
\item \code{"geobr"}: Uses \code{\link[geobr:read_municipal_seat]{read_municipal_seat()}}
from the \code{\link[geobr:geobr]{geobr}} package to retrieve the coordinates.
\item \code{"geocodebr"}: Uses the \code{\link[geocodebr:geocode]{geocode()}} from the
\code{\link[geocodebr:geocodebr]{geocodebr}} package to retrieve the coordinates.
}}

\item{force}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to force the download of the data again (default: \code{FALSE}).}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the following columns:
\itemize{
\item \code{region_code}: The region code.
\item \code{region}: The region name.
\item \code{state_code}: The state code.
\item \code{state}: The state name.
\item \code{federal_unit}: The state abbreviation.
\item \code{municipality_code}: The municipality code.
\item \code{municipality}: The municipality name.
\item \code{latitude}: The municipality latitude.
\item \code{longitude}: The municipality longitude.
}
}
\description{
\code{brazil_municipality()} returns a \code{\link[tibble:tibble]{tibble}} with data
about Brazilian municipalities.

This function normalizes names and objects from the
\code{\link[geobr:read_municipality]{read_municipality()}} function of the
\code{\link[geobr:geobr]{geobr}} package, adding latitude and longitude
coordinates for each municipality.

\strong{Note:} This function requires an internet connection to work and the
\href{https://ipeagit.github.io/geobr/}{\code{geobr}} or
\href{https://ipeagit.github.io/geocodebr/}{\code{geocodebr}} package to be
installed, depending on the chosen method for retrieving coordinates.
}
\details{
Data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} and \code{\link[geocodebr:geocodebr]{geocodebr}} R packages.

Both packages are produced by Brazil's Institute for Applied Economic
Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the Brazilian
Institute of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE})
data.
}
\examples{
library(curl)
library(dplyr)

\dontrun{
  if (has_internet()) {
    brazil_municipality() |> glimpse()
  }
}

\dontrun{
  if (has_internet()) {
    brazil_municipality(municipality = "Belém") |> glimpse()
  }
}

\dontrun{
  if (has_internet()) {
    brazil_municipality(municipality = "Belém", state = "Pará") |> glimpse()
  }
}

\dontrun{
  if (has_internet()) {
    brazil_municipality(municipality = c("Belém", "São Paulo")) |> glimpse()
  }
}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_fu}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region}()},
\code{\link{brazil_region_code}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_code}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
