% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_region.R
\name{brazil_region}
\alias{brazil_region}
\title{Get Brazilian regions}
\usage{
brazil_region(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names, abbreviations, or numeric codes of Brazilian states or federal
units. Region and municipality codes are also supported. If \code{NULL},
returns a vector with all Brazilian regions. (default: \code{NULL})}
}
\value{
A \code{\link[base:character]{character}} vector with the names of
Brazilian regions.
}
\description{
\code{brazil_region()} returns a vector with the names of
\href{https://en.wikipedia.org/wiki/Regions_of_Brazil}{Brazilian regions}.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
brazil_region()
#> [1] "North" "Northeast" "South" "Southeast" "Central-West" # Expected

brazil_region("sp")
#> [1] "Southeast" # Expected

brazil_region("sao paulo")
#> [1] "Southeast" # Expected

brazil_region(c(1, 4))
#> [1] "North" "South" # Expected

brazil_region(35) # State of São Paulo
#> [1] "Southeast" # Expected

brazil_region(3550308) # Municipality of São Paulo
#> [1] "Southeast" # Expected

brazil_region(35503081) # >7 digits
#> [1] NA # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_fu}()},
\code{\link{brazil_municipality}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region_code}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_code}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
