% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_region_code.R
\name{brazil_region_code}
\alias{brazil_region_code}
\title{Get Brazilian region codes}
\usage{
brazil_region_code(x = NULL)
}
\arguments{
\item{x}{(optional) A \code{\link[base:character]{character}} vector containing
the names of Brazilian regions, states, or federal units. If \code{NULL},
returns a named vector with all Brazilian region codes (default: \code{NULL}).}
}
\value{
An \code{\link[base:integer]{integer}} vector with the
\href{https://www.ibge.gov.br/}{IBGE} codes of Brazilian regions.
}
\description{
\code{brazil_region_code()} returns a vector with the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) codes for
Brazilian regions.
}
\details{
Data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}, n.d.).
}
\examples{
brazil_region_code()

brazil_region_code("north")
#> [1] 1 # Expected

brazil_region_code(c("north", "central-west"))
#> [1] 1 5 # Expected

brazil_region_code("sao paulo")
#> [1] 3 # Expected

brazil_region_code("sp")
#> [1] 3 # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_fu}()},
\code{\link{brazil_municipality}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_code}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
