% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_state.R
\name{brazil_state}
\alias{brazil_state}
\title{Get Brazilian state names}
\usage{
brazil_state(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names or numeric codes of Brazilian regions or federal units.
Municipality codes are also supported. If \code{NULL},
returns a \code{\link[base:character]{character}} vector with all Brazilian
state names (default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the names of
Brazilian states.
}
\description{
\code{brazil_state()} returns a vector with the names of Brazilian states.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
brazil_state()

brazil_state("rj")
#> [1] "Rio de Janeiro" # Expected

brazil_state("rio de janeiro")
#> [1] "Rio de Janeiro" # Expected

brazil_state(33)
#> [1] "Rio de Janeiro" # Expected

brazil_state(3302700) # Maricá
#> [1] "Rio de Janeiro" # Expected

brazil_state(33027001) # >7 digits
#> [1] NA # Expected

brazil_state(39027001) # Non-existent state code
#> [1] NA # Expected

brazil_state("southeast")
#> [1] "Espírito Santo" "Minas Gerais"   "Rio de Janeiro" # Expected
#> [4] "São Paulo"

brazil_state(3)
#> [1] "Espírito Santo" "Minas Gerais"   "Rio de Janeiro" # Expected
#> [4] "São Paulo"
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_fu}()},
\code{\link{brazil_municipality}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region}()},
\code{\link{brazil_region_code}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_code}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
