% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_state_code.R
\name{brazil_state_code}
\alias{brazil_state_code}
\title{Get Brazilian state codes}
\usage{
brazil_state_code(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names of Brazilian states or federal units. Municipality codes are also
supported. If \code{NULL}, returns a vector with all state codes
(default: \code{NULL}).}
}
\value{
An \code{\link[base:integer]{integer}} vector with the
\href{https://www.ibge.gov.br/}{IBGE} codes of Brazilian states.
}
\description{
\code{brazil_state_code()} returns a vector with the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) codes for
Brazilian states.
}
\details{
Data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}, n.d.).
}
\examples{
brazil_state_code()

brazil_state_code("ac")
#> [1] 12 # Expected

brazil_state_code("acre")
#> [1] 12 # Expected

brazil_state_code(3550308) # São Paulo
#> [1] 35 # Expected

brazil_state_code(35503081) # >7 digits
#> [1] NA # Expected

brazil_state_code(3912345) # Non-existent state code
#> [1] NA # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{brazil_fu}()},
\code{\link{brazil_municipality}()},
\code{\link{brazil_municipality_code}()},
\code{\link{brazil_municipality_coords}()},
\code{\link{brazil_municipality_latitude}()},
\code{\link{brazil_municipality_longitude}()},
\code{\link{brazil_region}()},
\code{\link{brazil_region_code}()},
\code{\link{brazil_render_address}()},
\code{\link{brazil_state}()},
\code{\link{brazil_state_by_utc}()},
\code{\link{brazil_state_capital}()},
\code{\link{brazil_state_latitude}()},
\code{\link{brazil_state_longitude}()}
}
\concept{Brazil functions}
