% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_geobr_year.R
\name{closest_geobr_year}
\alias{closest_geobr_year}
\title{Get the closest year available in the \code{geobr} package}
\usage{
closest_geobr_year(year, type = "country", verbose = TRUE)
}
\arguments{
\item{year}{An \code{\link[checkmate:checkIntegerish]{integerish}} vector with the
year to find the closest available year in the \code{geobr} package.}

\item{type}{(optional) A \code{\link[base:character]{character}} string indicating
the type of data to find the closest year for. It can be one of the
following: \code{"municipality"}, \code{"municipal_seat"}, \code{"state"}, or \code{"country"}
(default: \code{"country"}).}

\item{verbose}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to print a warning message if the specified year is not available
in the \code{geobr} package. Only applicable if \code{year} is a single value
(default: \code{TRUE}).}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the closest year available
in the geobr package for the specified type of data.
}
\description{
\code{closest_geobr_year()} returns the closest year available in the
\code{\link[geobr:geobr]{geobr}} package for a specified type of data.
}
\examples{
closest_geobr_year(2025, type = "municipality")
#> [1] 2024 # Expected

closest_geobr_year(2025, type = "municipal_seat")
#> [1] 2010 # Expected

closest_geobr_year(2025, type = "state")
#> [1] 2020 # Expected

closest_geobr_year(2025, type = "country")
#> [1] 2020 # Expected

closest_geobr_year(c(2025, 1999, NA, 1800), type = "country")
#> [1] 2020 2000   NA 1872 # Expected
}
\seealso{
Other utility functions: 
\code{\link{fix_postal_code}()},
\code{\link{map_fill_data}()},
\code{\link{remove_unique_outliers}()},
\code{\link{test_geobr_connection}()},
\code{\link{unique_outliers}()}
}
\concept{utility functions}
