% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_names.R
\name{country_names}
\alias{country_names}
\title{Get country names}
\usage{
country_names(format = "common name")
}
\arguments{
\item{format}{(optional) A \code{\link[base:character]{character}} string indicating
the format to be returned. The options are:
\itemize{
\item \code{"name"}: The name of the country.
\item \code{"common name"}: The common name of the country.
\item \code{"official name"}: The official name of the country.
\item \code{"alpha 2"}: The two-letter country code.
\item \code{"alpha 3"}: The three-letter country code.
\item \code{"numeric"}: The numeric country code.
}}
}
\value{
A \code{\link[base:character]{character}} vector with the names or codes of
all countries present in the
\href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166-1} standard.
}
\description{
\code{country_names()} returns a vector with the names or codes of all
countries present in the
\href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166-1} standard.

\strong{Note:} This function requires the
\href{https://CRAN.R-project.org/package=ISOcodes}{\code{ISOcodes}} package to be
installed.
}
\details{
The data from this function is based on data from the International
Organization for Standardization (ISO)
(\href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166-1}) via the
\code{\link[ISOcodes:ISO_3166]{ISOcodes}} R package. Click
\href{https://www.iso.org/iso-3166-country-codes.html}{here} to learn more.

The \code{\link[ISOcodes:ISO_3166]{ISOcodes}} R package uses XML files provided by
Debian's \href{https://salsa.debian.org/iso-codes-team/iso-codes}{\code{iso-codes}}
package for the data.
}
\examples{
country_names(format = "alpha 3")
}
\concept{World functions}
