% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_points_on_land.R
\name{filter_points_on_land}
\alias{filter_points_on_land}
\title{Filter points on land}
\usage{
filter_points_on_land(data, geometry)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble}} with the data points to be filtered.
It must have columns \code{longitude} and \code{latitude}.}

\item{geometry}{A \code{\link[sf:sf]{sf}} object with the geometry to be used for
filtering.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
\code{filter_points_on_land()} filters latitude/longitude points that intersects
with a given geometry.
}
\examples{
library(dplyr)
library(geobr)
library(sf)

data <- tibble(
  latitude = get_brazil_state_latitude(),
  longitude = get_brazil_state_longitude()
)

data

geometry <- read_state(code = "SP") |> pull(geom)

geometry |> st_bbox()

filter_points_on_land(data, geometry)
}
\concept{\code{sf} functions}
