% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_address_by_postal_code.R
\name{get_brazil_address_by_postal_code}
\alias{get_brazil_address_by_postal_code}
\title{Get a Brazilian address by its postal code via reverse geocoding}
\usage{
get_brazil_address_by_postal_code(
  postal_code,
  method = "qualocep",
  fix_code = TRUE,
  limit = 10
)
}
\arguments{
\item{postal_code}{A \code{\link[base:character]{character}} vector with the postal
code(s) to be used to. The postal code must be in the format \code{XXXXXXXX},
where \code{X} is a digit.}

\item{method}{A \code{\link[base:character]{character}} value indicating the method
to be used to retrieve the address data. The available options are \code{"osm"},
\code{"google"},\code{"qualocep"}, and \code{"viacep"} (Default: \code{qualocep}).}

\item{fix_code}{A \code{\link[base:logical]{logical}} flag indicating if the postal
code must be fixed before being used (Default: \code{TRUE}).}

\item{limit}{A \code{\link[base:numeric]{numeric}} value indicating the maximum
number of results to return. If the value is \code{Inf}, all postal codes will
be used (Default: \code{10}).}
}
\value{
A \code{\link[dplyr:reexports]{tibble}} with the following columns:
\itemize{
\item \code{postal_code}: A \code{\link[base:character]{character}} vector with the postal
codes.
\item \code{street}: A \code{\link[base:character]{character}} vector with the full names of
the streets.
\item \code{complement}: A \code{\link[base:character]{character}} vector with the
complements of the addresses.
\item \code{neighborhood}: A \code{\link[base:character]{character}} vector with the
neighborhoods.
\item \code{municipality_code}: An \code{\link[base:integer]{integer}} vector with the codes
of the Brazilian Institute of Geography and Statistics
(\href{https://www.ibge.gov.br/}{IBGE}) for Brazilian municipalities.
\item \code{municipality}: A \code{\link[base:character]{character}} vector with the names of
the municipalities.
\item \code{state_code}: An \code{\link[base:integer]{integer}} vector with the codes
of the Brazilian Institute of Geography and Statistics
(\href{https://www.ibge.gov.br/}{IBGE}) for the Brazilian states.
\item \code{state}: A \code{\link[base:character]{character}} vector with the names of the
states.
\item \code{region}: A \code{\link[base:character]{character}} vector with the
Brazilian regions.
\item \code{address}: A \code{\link[base:character]{character}} vector with the
full addresses.
\item \code{latitude}: A \code{\link[base:numeric]{numeric}} vector with the latitude values
of the postal codes.
\item \code{longitude}: A \code{\link[base:numeric]{numeric}} vector with the longitude
values of the postal codes.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{get_brazil_address_by_postal_code()} retrieves a Brazilian address based
on a postal code using
\href{https://en.wikipedia.org/wiki/Reverse_geocoding}{reverse geocoding}.

Please note that the accuracy of the results may vary depending on the
method used.
}
\details{
The source of the data will depend on the method used. Run
\code{\link[tidygeocoder:geo]{?tidygeocoder::geo}} to learn more. The only
exception is the \code{"qualocep"} and \code{"viacep"} method. These methods will
return the address data from the \href{https://www.qualocep.com/}{Qual o CEP}
database and the \href{https://viacep.com.br/}{ViaCEP} API, respectively.

See \code{\link[tidygeocoder:geo]{?tidygeocoder::geo}} to learn how to use
\verb{method = "google}.
}
\examples{
\dontrun{
  "01014908" |>
    get_brazil_address_by_postal_code(method = "osm") |>
    dplyr::glimpse()

  "01014908" |>
    get_brazil_address_by_postal_code(method = "google") |>
    dplyr::glimpse()

  "01014908" |>
    get_brazil_address_by_postal_code(method = "qualocep") |>
    dplyr::glimpse()

  "01014908" |>
    get_brazil_address_by_postal_code(method = "viacep") |>
    dplyr::glimpse()

  c("01014-908", NA, "05650-905") |>
    get_brazil_address_by_postal_code(method = "qualocep") |>
    dplyr::glimpse()
}
}
\seealso{
Other API functions: 
\code{\link{get_qualocep_data}()}
}
\concept{API functions}
