% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_municipality.R
\name{get_brazil_municipality}
\alias{get_brazil_municipality}
\title{Get Brazilian municipalities data}
\usage{
get_brazil_municipality(
  municipality = NULL,
  state = NULL,
  year = 2017,
  force = FALSE
)
}
\arguments{
\item{municipality}{(Optional) A \code{\link[base:character]{character}} vector
with the name of the municipalities. If \code{NULL} the function returns all
municipalities (Default: \code{NULL}).}

\item{state}{(Optional) A \code{\link[base:character]{character}} vector with the
name of the states (Default: \code{NULL}).}

\item{year}{(Optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(Default: \code{2017}).}

\item{force}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
whether to force the download of the data again (Default: \code{FALSE}).}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the following columns:
\itemize{
\item \code{country}: The country name.
\item \code{region_code}: The region code.
\item \code{region}: The region name.
\item \code{state_code}: The state code.
\item \code{state}: The state name.
\item \code{federal_unit}: The state abbreviation.
\item \code{municipality_code}: The municipality code.
\item \code{municipality}: The municipality name.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_municipality()} returns a \code{\link[tibble:tibble]{tibble}} with data
about Brazilian municipalities.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} R package.

The \code{\link[geobr:geobr]{geobr}} package is produced by Brazil's Institute for
Applied Economic Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the
Brazilian Institute of Geography and Statistics (IBGE) data. You can see a
list of all \code{\link[geobr:geobr]{geobr}} datasets by running
\code{\link[geobr:list_geobr]{geobr::list_geobr()}}.
}
\examples{
get_brazil_municipality() |> dplyr::glimpse()

get_brazil_municipality(municipality = "Belém")

get_brazil_municipality(municipality = "Belém", state = "Pará")

get_brazil_municipality(municipality = c("Belém", "São Paulo"))
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
