% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_municipality_code.R
\name{get_brazil_municipality_code}
\alias{get_brazil_municipality_code}
\title{Get Brazilian municipalities codes}
\usage{
get_brazil_municipality_code(
  municipality,
  state = NULL,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  names = TRUE,
  ...
)
}
\arguments{
\item{municipality}{(optional) A \code{\link[base:character]{character}} vector
with the name of the municipalities. If \code{NULL} the function returns all
municipalities (default: \code{NULL}).}

\item{state}{(optional) A \code{\link[base:character]{character}} vector with the
name of the states (default: \code{NULL}).}

\item{year}{(optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(default: \code{Sys.Date() |> substr(1, 4) |> as.numeric()}).}

\item{names}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to return the names of the municipalities as names of the
vector (default: \code{TRUE}).}

\item{...}{(optional) Additional arguments passed to
\code{\link[=get_brazil_municipality]{get_brazil_municipality()}}.}
}
\value{
An \code{\link[base:integer]{integer}} vector with the IBGE codes of
Brazilian municipalities.
}
\description{
\code{get_brazil_municipality()} returns a vector with codes of the Brazilian
Institute of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) for
Brazilian municipalities.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} R package.

The \code{\link[geobr:geobr]{geobr}} package is produced by Brazil's Institute for
Applied Economic Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the
Brazilian Institute of Geography and Statistics (IBGE) data. You can see a
list of all \code{\link[geobr:geobr]{geobr}} datasets by running
\code{\link[geobr:list_geobr]{geobr::list_geobr()}}.
}
\examples{
\dontrun{
  get_brazil_municipality_code(municipality = "Belém")

  get_brazil_municipality_code(municipality = "Belém", names = FALSE)

  get_brazil_municipality_code(municipality = "Belém", state = "Pará")

  get_brazil_municipality_code(c("Rio de Janeiro", "São Paulo"))
}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_coords}()},
\code{\link{get_brazil_municipality_latitude}()},
\code{\link{get_brazil_municipality_longitude}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
