% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_municipality_coords.R
\name{get_brazil_municipality_coords}
\alias{get_brazil_municipality_coords}
\title{Get Brazilian municipalities geographic coordinates}
\usage{
get_brazil_municipality_coords(
  municipality_code = NULL,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  coords_method = "geobr",
  force = FALSE
)
}
\arguments{
\item{municipality_code}{(optional) An
\code{\link[checkmate:checkIntegerish]{integerish}} vector with the IBGE codes of
Brazilian municipalities. Use
\code{\link[=get_brazil_municipality_code]{get_brazil_municipality_code()}} to
obtain codes from municipality names and states. If \code{NULL} the function
returns all municipalities (default: \code{NULL}).}

\item{year}{(optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(default: \code{Sys.Date() |> substr(1, 4) |> as.numeric()}).}

\item{coords_method}{(optional) A string indicating the method to retrieve
the latitude and longitude coordinates of the municipalities. Options are:
\itemize{
\item \code{"geobr"}: Uses \code{\link[geobr:read_municipal_seat]{read_municipal_seat()}}
from the \code{\link[geobr:geobr]{geobr}} package to retrieve the coordinates.
\item \code{"geocodebr"}: Uses the \code{\link[geocodebr:geocode]{geocode()}} from the
\code{\link[geocodebr:geocodebr]{geocodebr}} package to retrieve the coordinates.
(default: \code{"geobr"}).
}}

\item{force}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to force the download of the data again (default: \code{FALSE}).}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the following columns:
\itemize{
\item \code{municipality_code}: The municipality code.
\item \code{latitude}: The municipality latitude.
\item \code{longitude}: The municipality longitude.
}
}
\description{
\code{get_brazil_municipality_coords()} returns a \code{\link[tibble:tibble]{tibble}}
with the latitude and longitude coordinates of Brazilian municipalities.

\strong{Note:} This function requires an internet connection to work and the
\href{https://ipeagit.github.io/geobr/}{\code{geobr}} or
\href{https://ipeagit.github.io/geocodebr/}{\code{geocodebr}} package to be
installed, depending on the chosen method for retrieving coordinates.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} and \code{\link[geocodebr:geocodebr]{geocodebr}} R packages.

Both packages are produced by Brazil's Institute for Applied Economic
Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the Brazilian
Institute of Geography and Statistics (IBGE) data.
}
\examples{
\dontrun{
  get_brazil_municipality_coords() |> dplyr::glimpse()

  get_brazil_municipality_coords(municipality_code = 3550308)

  get_brazil_municipality_coords(municipality_code = 3550)

  get_brazil_municipality_coords(municipality_code = c(3550308, 3304557))
}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_municipality_latitude}()},
\code{\link{get_brazil_municipality_longitude}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
