% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_municipality_latitude.R
\name{get_brazil_municipality_latitude}
\alias{get_brazil_municipality_latitude}
\title{Get Brazilian municipalities latitude}
\usage{
get_brazil_municipality_latitude(
  municipality_code,
  year = lubridate::year(Sys.Date()),
  names = TRUE,
  ...
)
}
\arguments{
\item{municipality_code}{An \code{\link[checkmate:checkIntegerish]{integerish}} vector
with the IBGE codes of Brazilian municipalities. Use
\code{\link[=get_brazil_municipality_code]{get_brazil_municipality_code()}}
to obtain codes from municipality names and states.}

\item{year}{(optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(default: \code{Sys.Date() |> lubridate::year()}).}

\item{names}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to return the names of the municipalities as names of the
vector (default: \code{TRUE}).}

\item{...}{(optional) Additional arguments passed to
\code{\link[=get_brazil_municipality]{get_brazil_municipality()}}.}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the latitude of the
Brazilian municipalities.
}
\description{
\code{get_brazil_municipality_latitude()} returns the latitude of Brazilian
municipalities.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} R package.

The \code{\link[geobr:geobr]{geobr}} package is produced by Brazil's Institute for
Applied Economic Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the
Brazilian Institute of Geography and Statistics (IBGE) data. You can see a
list of all \code{\link[geobr:geobr]{geobr}} datasets by running
\code{\link[geobr:list_geobr]{geobr::list_geobr()}}.
}
\examples{
get_brazil_municipality_latitude(3550308)

get_brazil_municipality_latitude(c(3550308, 3500204))

get_brazil_municipality_latitude(c(3550308, 1000, 3500204))

get_brazil_municipality_latitude(c(3550308, NA, 3500204))
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_municipality_coords}()},
\code{\link{get_brazil_municipality_longitude}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
