% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_municipality_longitude.R
\name{get_brazil_municipality_longitude}
\alias{get_brazil_municipality_longitude}
\title{Get Brazilian municipalities longitude}
\usage{
get_brazil_municipality_longitude(
  municipality_code,
  year = lubridate::year(Sys.Date()),
  coords_method = "geobr",
  names = TRUE,
  ...
)
}
\arguments{
\item{municipality_code}{An \code{\link[checkmate:checkIntegerish]{integerish}} vector
with the IBGE codes of Brazilian municipalities. Use
\code{\link[=get_brazil_municipality_code]{get_brazil_municipality_code()}}
to obtain codes from municipality names and states.}

\item{year}{(optional) An \code{\link[checkmate:checkInt]{integerish}} number
indicating the year of the data regarding the municipalities
(default: \code{Sys.Date() |> lubridate::year()}).}

\item{coords_method}{(optional) A string indicating the method to retrieve
the latitude and longitude coordinates of the municipalities. Options are:
\itemize{
\item \code{"geobr"}: Uses \code{\link[geobr:read_municipal_seat]{read_municipal_seat()}}
from the \code{\link[geobr:geobr]{geobr}} package to retrieve the coordinates.
\item \code{"geocodebr"}: Uses the \code{\link[geocodebr:geocode]{geocode()}} from the
\code{\link[geocodebr:geocodebr]{geocodebr}} package to retrieve the coordinates.
(default: \code{"geobr"}).
}}

\item{names}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to return the names of the municipalities as names of the
vector (default: \code{TRUE}).}

\item{...}{(optional) Additional arguments passed to
\code{\link[=get_brazil_municipality]{get_brazil_municipality()}}.}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the longitude of the
Brazilian municipalities.
}
\description{
\code{get_brazil_municipality_longitude()} returns the longitude of Brazilian
municipalities.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) via the
\code{\link[geobr:geobr]{geobr}} R package.

The \code{\link[geobr:geobr]{geobr}} package is produced by Brazil's Institute for
Applied Economic Research (\href{https://www.ipea.gov.br/}{IPEA}) and access the
Brazilian Institute of Geography and Statistics (IBGE) data. You can see a
list of all \code{\link[geobr:geobr]{geobr}} datasets by running
\code{\link[geobr:list_geobr]{geobr::list_geobr()}}.
}
\examples{
get_brazil_municipality_longitude(3550308)

get_brazil_municipality_longitude(c(3550308, 3500204))

get_brazil_municipality_longitude(c(3550308, 1000, 3500204))

get_brazil_municipality_longitude(c(3550308, NA, 3500204))
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_municipality_coords}()},
\code{\link{get_brazil_municipality_latitude}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
