% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_region_code.R
\name{get_brazil_region_code}
\alias{get_brazil_region_code}
\title{Get Brazilian region codes}
\usage{
get_brazil_region_code(x = NULL)
}
\arguments{
\item{x}{(Optional) A \code{\link[base:character]{character}} vector with the names
of Brazilian regions. If \code{NULL}, returns all Brazilian region codes
(Default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the Brazilian
region codes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_region_code()} returns a vector with the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) codes for
Brazilian regions.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
get_brazil_region_code()
#> [1] 1 2 3 4 5 # Expected

get_brazil_region_code("north")
#> [1] 1 # Expected

get_brazil_region_code(c("north", "central-west"))
#> [1] 1 5 # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
