% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_by_utc.R
\name{get_brazil_state_by_utc}
\alias{get_brazil_state_by_utc}
\title{Get Brazilian state names by UTC}
\usage{
get_brazil_state_by_utc(utc = -3, type = "fu")
}
\arguments{
\item{utc}{(Optional) An \code{\link[checkmate:checkInt]{integerish}} number
with the UTC offset. Available choices are \code{-5}, \code{-4}, \code{-3}, or \code{-2}
(Default: \code{-3}).}

\item{type}{(Optional) A \code{\link[base:character]{character}} string specifying
the type of value to return. Available choices are \code{"state"} or \code{"fu"}
(Default: \code{"fu"}).}
}
\value{
A \code{\link[base:character]{character}} vector with the names of
Brazilian states or abbreviations of Brazilian federal units.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_state_by_utc()} returns a vector with the names of
Brazilian states or abbreviations of Brazilian federal units by the
\href{https://en.wikipedia.org/wiki/Coordinated_Universal_Time}{UTC}
offset.
}
\details{
The data from this function is based on the 2024b dataset
(Released 2024-09-04) from the Internet Assigned Numbers Authority
(IANA, 2024)
}
\examples{
get_brazil_state_by_utc(-3, type = "fu")

get_brazil_state_by_utc(-3, type = "state")
}
\references{
Internet Assigned Numbers Authority. (2024). \emph{Time zone database
(No. 2024b)} [Dataset]. \url{https://www.iana.org/time-zones}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()}
}
\concept{Brazil functions}
