% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_capital.R
\name{get_brazil_state_capital}
\alias{get_brazil_state_capital}
\title{Get Brazilian state capital names}
\usage{
get_brazil_state_capital(x = NULL)
}
\arguments{
\item{x}{(Optional) A \code{\link[base:character]{character}} vector with the names
of Brazilian states or federal units. If \code{NULL}, returns all state capital
names (Default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the names of
Brazilian state capitals.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_state_capital()} returns a vector with the capital names of
Brazilian states or federal units.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
get_brazil_state_capital()

get_brazil_state_capital("piaui")
#> [1] "Teresina" # Expected

get_brazil_state_capital("pi")
#> [1] "Teresina" # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()}
}
\concept{Brazil functions}
