% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_latitude.R
\name{get_brazil_state_latitude}
\alias{get_brazil_state_latitude}
\title{Get Brazilian state latitude}
\usage{
get_brazil_state_latitude(x = NULL)
}
\arguments{
\item{x}{(Optional) A \code{\link[base:character]{character}} vector with the names
of Brazilian states or federal units. If \code{NULL}, returns all state
latitudes (Default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the latitude of
Brazilian states.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_state_latitude()} returns a vector with the latitude of
Brazilian state capitals.
}
\details{
The data from this function is based on Google's Geocoding API gathered via
the \code{\link{tidygeocoder}} R package.
}
\examples{
get_brazil_state_latitude()

get_brazil_state_latitude("sao paulo")
#> [1] -23.55065 # Expected

get_brazil_state_latitude("sp")
#> [1] -23.55065 # Expected
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_longitude}()}
}
\concept{Brazil functions}
