% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_longitude.R
\name{get_brazil_state_longitude}
\alias{get_brazil_state_longitude}
\title{Get Brazilian state longitude}
\usage{
get_brazil_state_longitude(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names of Brazilian states or federal units. Municipality and state codes
are also supported. If \code{NULL}, returns a vector with all state longitudes
(default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the longitude of
Brazilian state capitals.
}
\description{
\code{get_brazil_state_longitude()} returns a vector with the longitude of
Brazilian state capitals.
}
\details{
The data from this function is based on Google's Geocoding API gathered via
the \code{\link[tidygeocoder:tidygeocoder-package]{tidygeocoder}} R package.
}
\examples{
get_brazil_state_longitude()

get_brazil_state_longitude("sp")
#> [1] -46.63338 # Expected

get_brazil_state_longitude("sao paulo")
#> [1] -46.63338 # Expected

get_brazil_state_longitude(35) # State of São Paulo
#> [1] -46.63338 # Expected

get_brazil_state_longitude(3550308) # Municipality of São Paulo
#> [1] -46.63338 # Expected

get_brazil_state_longitude(35503081) # >7 digits
#> [1] NA # Expected

get_brazil_state_longitude(3912345) # Non-existent state code
#> [1] NA # Expected
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_municipality_coords}()},
\code{\link{get_brazil_municipality_latitude}()},
\code{\link{get_brazil_municipality_longitude}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
