% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_fill_data.R
\name{get_map_fill_data}
\alias{get_map_fill_data}
\title{Get data to fill a map}
\usage{
get_map_fill_data(
  data,
  col_fill = NULL,
  col_code,
  name_col_value = "n",
  name_col_ref = col_code,
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble}} with the data to be used as fill.}

\item{col_fill}{(Optional) A \code{\link[base:character]{character}} string with
the column name to be used as fill. If \code{NULL}, the function will count
the number of occurrences of each value in \code{col_code} (Default: \code{NULL}).}

\item{col_code}{A \code{\link[base:character]{character}} string with the column
name to be used as reference.}

\item{name_col_value}{(Optional) A \code{\link[base:character]{character}} string
with the name of the column to be used as value (Default: \code{"n"}).}

\item{name_col_ref}{(Optional) A \code{\link[base:character]{character}} string
with the name of the column to be used as reference (Default: \code{col_code}).}

\item{quiet}{(Optional) A \code{\link[base:logical]{logical}} flag to suppress
messages (Default: \code{FALSE}).}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{get_map_fill_data()} prepares data to be used as fill in a map.
}
\examples{
library(dplyr)

data <- tibble(
  state = c("SP", "RJ", "MG", "SP", "RJ", "MG"),
  value = c(1, 2, 3, 4, 5, 6)
)

get_map_fill_data(data, col_fill = NULL, col_code = "state")

get_map_fill_data(data, col_fill = "value", col_code = "state")
}
\seealso{
Other utility functions: 
\code{\link{fix_postal_code}()}
}
\concept{utility functions}
