% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sidra_by_year.R
\name{get_sidra_by_year}
\alias{get_sidra_by_year}
\title{Get and aggregate data by year from SIDRA API}
\usage{
get_sidra_by_year(years, api_start, api_end)
}
\arguments{
\item{years}{An \code{\link[checkmate:checkIntegerish]{integerish}} vector with the
years to download.}

\item{api_start}{A string specifying the initial part of the SIDRA API URL,
up to (but not including) the year segment. See the Details section for
guidance.}

\item{api_end}{A string specifying the final part of the SIDRA API URL,
immediately following the year segment. See the Details section for
guidance.}
}
\value{
A \link[tibble:tibble]{tibble} containing the combined data for all
requested years, as retrieved from the SIDRA API.
}
\description{
\code{get_sidra_by_year()} retrieves data from the Brazilian Institute of
Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) Automatic
Retrieval System (\href{https://sidra.ibge.gov.br/}{SIDRA}) API for a specified
range of years.

This function addresses the SIDRA API's limitations on the volume of data
that can be downloaded in a single request. It downloads data for each year
individually and then combines the results into a single tibble.

\strong{Note:} This function requires an active internet connection and the
\href{https://CRAN.R-project.org/package=sidrar}{\code{sidrar}} package to be
installed.
}
\details{
To construct the API call, follow these steps:
\enumerate{
\item Visit the SIDRA website.
\item Locate the desired table containing your data.
\item Configure the parameters for data retrieval (e.g., variable, sex, years).
\item Click the share button (link symbol) at the end of the page.
\item If a checkbox labeled "Usar períodos relativos, quando possível." appears,
uncheck it, reload the page, and click the share button again.
\item Copy the portion of the "Parâmetros para a API" URL that starts with "/t"
(e.g., /t/6407...).
}

You need to provide the function with separate parts of the API URL. For
example:

\if{html}{\out{<div class="sourceCode text">}}\preformatted{|-------- Start ------|--- Years ----|----- End -----|
/t/6407/n6/all/v/606/p/2021,2022,2023/c2/6794/c58/1140
}\if{html}{\out{</div>}}

If you have difficulty identifying the correct segments, try adjusting the
table settings, selecting different years, and examining how the URL changes.
}
\examples{
\dontrun{
  get_sidra_by_year(
    years = 2010:2011,
    api_start = "/t/1612/n6/all/v/109/p/",
    api_end = "/c81/2692"
  )
}
}
\concept{other functions.}
