% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fill_data.R
\name{map_fill_data}
\alias{map_fill_data}
\title{Prepare data to fill a map}
\usage{
map_fill_data(
  data,
  col_fill = NULL,
  col_code,
  name_col_value = "n",
  name_col_ref = col_code,
  quiet = FALSE
)
}
\arguments{
\item{data}{A \code{\link[tibble:tibble]{tibble}} with the data to be used as fill.}

\item{col_fill}{(optional) A \code{\link[base:character]{character}} string with
the column name to be used as fill. If \code{NULL}, the function will count
the number of occurrences of each value in \code{col_code} (default: \code{NULL}).}

\item{col_code}{A \code{\link[base:character]{character}} string with the column
name to be used as reference.}

\item{name_col_value}{(optional) A \code{\link[base:character]{character}} string
with the name of the column to be used as value (default: \code{"n"}).}

\item{name_col_ref}{(optional) A \code{\link[base:character]{character}} string
with the name of the column to be used as reference (default: \code{col_code}).}

\item{quiet}{(optional) A \code{\link[base:logical]{logical}} flag to suppress
messages (default: \code{FALSE}).}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
\code{map_fill_data()} prepares data to be used as fill in a map.
}
\examples{
\dontrun{
  library(curl)
  library(dplyr)
  library(geodata)
  library(ggplot2)
  library(terra)
  library(tidyterra)

  if (has_internet()) {
    brazil_states <- geodata::gadm("BRA", level = 1, path = tempdir())

    brazil_states |>
      ggplot() +
      geom_spatvector(fill = "#5F45BF", color = "white")

    data <- tibble(
      state = sample(brazil_states$NAME_1, size = 1000, replace = TRUE),
      value = sample(1:1000, size = 1000, replace = TRUE)
    )

    data

    data <- map_fill_data(data, col_fill = "value", col_code = "state")

    data

    left_join(brazil_states, data, by = c("NAME_1" = "state")) |>
      ggplot() +
      geom_spatvector(aes(fill = n), color = "white") +
      scale_fill_continuous(
          palette = c("#072359", "#3243A6", "#9483AF"),
          na.value = "white"
      ) +
      labs(fill = NULL)
  }
}
}
\seealso{
Other utility functions: 
\code{\link{closest_geobr_year}()},
\code{\link{fix_postal_code}()},
\code{\link{remove_unique_outliers}()},
\code{\link{unique_outliers}()}
}
\concept{utility functions}
