% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{orb_coord_cartesian}
\alias{orb_coord_cartesian}
\title{Cartesian coordinates}
\usage{
orb_coord_cartesian(ratio = NULL)
}
\arguments{
\item{ratio}{Data units per y unit relative to x, or \code{NULL} for free.}
}
\value{
An object of class \code{orb_coord}.
}
\description{
The default coordinate system. Use it to return to Cartesian coordinates or
to force a fixed aspect ratio.
}
\examples{
orb(mtcars, x = wt, y = mpg) + orb_points() + orb_coord_cartesian()
}
