% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{orb_geo_points}
\alias{orb_geo_points}
\title{Point layer positioned by longitude and latitude}
\usage{
orb_geo_points(
  ...,
  data = NULL,
  size = 4,
  colour = "#F76707",
  alpha = 0.9,
  stroke = 0.8
)
}
\arguments{
\item{...}{Aesthetic mappings for this layer only, as \code{channel = column}.}

\item{data}{Optional data frame overriding the plot data.}

\item{size}{Point radius in pixels when \code{size} is not mapped (default 3.2).}

\item{colour}{Fixed colour used when the \code{colour} channel is not mapped.}

\item{alpha}{Opacity between 0 and 1.}

\item{stroke}{Outline width in pixels.}
}
\value{
An object of class \code{orb_layer}.
}
\description{
Places points on a map using the active projection. Use inside a plot that
also has an \code{\link[=orb_map]{orb_map()}} layer, mapping \code{x} to longitude and \code{y} to latitude.
}
\examples{
cities <- data.frame(long = c(2.35, -74, 151.2), lat = c(48.86, 40.7, -33.87),
                     pop = c(11, 19, 5))
orb(cities, x = long, y = lat, size = pop) +
  orb_map() + orb_points(colour = "#F76707") + orb_coord_map("robinson")
}
