% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{orb_labs}
\alias{orb_labs}
\title{Axis, legend and plot labels}
\usage{
orb_labs(
  title = NULL,
  subtitle = NULL,
  x = NULL,
  y = NULL,
  colour = NULL,
  fill = NULL,
  size = NULL,
  caption = NULL
)
}
\arguments{
\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{x, y}{Axis titles.}

\item{colour, fill, size}{Legend titles for those channels.}

\item{caption}{Caption shown under the plot.}
}
\value{
An object of class \code{orb_labs}, to be added to a plot.
}
\description{
Axis, legend and plot labels
}
\examples{
orb(mtcars, x = wt, y = mpg) + orb_points() +
  orb_labs(title = "Fuel use", x = "Weight", y = "Miles per gallon")
}
