% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{orb_options}
\alias{orb_options}
\title{Plot-level options}
\usage{
orb_options(legend = NULL, grid = NULL, interactive = NULL)
}
\arguments{
\item{legend}{Should a legend be drawn when a colour, fill or size mapping
is present? Default \code{TRUE}.}

\item{grid}{One of \code{"both"}, \code{"x"}, \code{"y"} or \code{"none"}.}

\item{interactive}{Should the SVG output embed interactive behaviour
(tooltips, hover highlighting, zoom, pan and legend toggling)? Default
\code{TRUE}. Raster output is unaffected.}
}
\value{
An object of class \code{orb_opts}, to be added to a plot.
}
\description{
Plot-level options
}
\examples{
orb(mtcars, x = wt, y = mpg) + orb_points() + orb_options(grid = "y")
}
