% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{+.orb_spec}
\alias{+.orb_spec}
\title{Add layers, scales, labels and themes to a plot}
\usage{
\method{+}{orb_spec}(e1, e2)
}
\arguments{
\item{e1}{An \code{orb_spec} object.}

\item{e2}{A layer, scale, theme, labels or coordinate object to add.}
}
\value{
An \code{orb_spec} object.
}
\description{
The \code{+} operator assembles a plot from parts, in the tradition of a layered
grammar of graphics.
}
\examples{
orb(mtcars, x = wt, y = mpg) + orb_points() + orb_theme_dark()
}
