% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unique_outliers.R
\name{remove_unique_outliers}
\alias{remove_unique_outliers}
\title{Remove unique outliers from raster files}
\usage{
remove_unique_outliers(file, n_iqr = 1.5)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector with the file paths of
the raster files to be processed. Supported file formats are
\href{https://en.wikipedia.org/wiki/GeoTIFF}{GeoTIFF}
(\code{.tif} or \code{.tiff}) and
\href{https://en.wikipedia.org/wiki/Esri_grid}{Esri ASCII} raster (\code{.asc}).}

\item{n_iqr}{(optional) A number specifying the multiplier of the
interquartile range
(\href{https://en.wikipedia.org/wiki/Interquartile_range}{IQR})
to define outliers. See \code{\link[=unique_outliers]{unique_outliers()}}
to learn more (default: \code{1.5}).}
}
\value{
An invisible \code{NULL} value. This function is used for its side
effects.
}
\description{
\code{remove_unique_outliers()} removes unique outliers from raster files
(\href{https://en.wikipedia.org/wiki/GeoTIFF}{GeoTIFF} or
\href{https://en.wikipedia.org/wiki/Esri_grid}{Esri ASCII} raster format) based
on the interquartile range
(\href{https://en.wikipedia.org/wiki/Interquartile_range}{IQR}).

This function processes each raster file by reading its values, identifying
unique outliers using the \code{\link[=unique_outliers]{unique_outliers()}} function,
and replacing those outlier values with \code{NA}. The modified raster is then
saved back to the same file, effectively overwriting the original data.

\strong{Note:} This function requires the \href{https://fs.r-lib.org/}{\code{fs}} package to
be installed.
}
\examples{
# Set the Environment -----

library(readr)
library(terra)

# Create a Fictional Esri ASCII File -----

asc_content <- c(
  "ncols         5",
  "nrows         5",
  "xllcorner     0.0",
  "yllcorner     0.0",
  "cellsize      1.0",
  "NODATA_value  -9999",
  "1 2 3 4 5 ",
  "6 7 8 9 10 ",
  "11 12 1000 14 15 ", # Extreme outlier (1000)
  "16 1 18 19 20 ",
  "21 22 23 24 25 "
)

temp_file <- tempfile(fileext = ".asc")

asc_content |> write_lines(temp_file)

# Visualize Values Before `remove_unique_outliers()` -----

temp_file |> rast() |> values(mat = FALSE)

# Visualize Values After `remove_unique_outliers()` -----

temp_file |> remove_unique_outliers()

temp_file |> rast() |> values(mat = FALSE)
}
\seealso{
Other utility functions: 
\code{\link{closest_geobr_year}()},
\code{\link{fix_postal_code}()},
\code{\link{map_fill_data}()},
\code{\link{test_geobr_connection}()},
\code{\link{unique_outliers}()}
}
\concept{utility functions}
