% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_date_line.R
\name{test_date_line}
\alias{test_date_line}
\title{Test if a spatial object crosses the international date line}
\usage{
test_date_line(x)
}
\arguments{
\item{x}{A \code{\link[sf:st_as_sf]{sf}}, \code{\link[terra:vect]{SpatVector}}, or
\code{\link[terra:rast]{SpatRaster}} object to be tested.}
}
\value{
A \code{\link[base:logical]{logical}} flag indicating if the object crosses
the international date line.
}
\description{
\code{test_date_line()} checks whether a given spatial object crosses the
international date line
(\href{https://en.wikipedia.org/wiki/International_Date_Line}{IDL}) by examining
its longitude extent.
}
\examples{
\dontrun{
   library(geodata)

   brazil_shape <- gadm(country = "BRA", level = 0)
   russia_shape <- gadm(country = "RUS", level = 0)

   test_date_line(brazil_shape)
   test_date_line(russia_shape)
}
}
\seealso{
Other utility functions: 
\code{\link{closest_geobr_year}()},
\code{\link{fix_postal_code}()},
\code{\link{remove_unique_outliers}()},
\code{\link{test_geobr_connection}()},
\code{\link{unique_outliers}()}
}
\concept{utility functions}
