% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_geobr_connection.R
\name{test_geobr_connection}
\alias{test_geobr_connection}
\title{Test \code{geobr} Package Connection with Its Server}
\usage{
test_geobr_connection()
}
\value{
A \code{\link[base:logical]{logical}} flag indicating if the connection
was successful.
}
\description{
\code{test_geobr_connection()} tests if the \code{geobr} package can successfully
connect to its server and retrieve data.

\strong{Note:} This function requires an active internet connection and the
\href{https://ipeagit.github.io/geobr/}{\code{geobr}} package to be installed.
}
\examples{
\dontrun{
  library(curl)
  library(geobr)

  if (has_internet()) {
    test_geobr_connection()
  }
}
}
\seealso{
Other utility functions: 
\code{\link{closest_geobr_year}()},
\code{\link{fix_postal_code}()},
\code{\link{map_fill_data}()},
\code{\link{remove_unique_outliers}()},
\code{\link{unique_outliers}()}
}
\concept{utility functions}
