% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_outliers.R
\name{unique_outliers}
\alias{unique_outliers}
\title{Return unique outliers}
\usage{
unique_outliers(x, n_iqr = 1.5)
}
\arguments{
\item{x}{An \code{\link[base:numeric]{numeric}} vector with at least 4 values.}

\item{n_iqr}{(optional) A number specifying the multiplier of the
interquartile range
(\href{https://en.wikipedia.org/wiki/Interquartile_range}{IQR})
to define outliers (default: \code{1.5}).}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the outliers of \code{x}.
}
\description{
\code{unique_outliers()} returns the unique outliers of a
\code{\link[base:numeric]{numeric}} vector based on the interquartile range
(\href{https://en.wikipedia.org/wiki/Interquartile_range}{IQR}).

This function first removes duplicated values from the input vector \code{x}
to ensure that outlier detection is based on unique values. It then
calculates the first (Q1) and third (Q3) quartiles, as well as the IQR.
Outliers are defined as values that fall below \code{Q1 - n_iqr * IQR} or above
\code{Q3 + n_iqr * IQR}, where \code{n_iqr} is a user-defined multiplier
(default is 1.5).
}
\examples{
c(1:10) |> unique_outliers()
#> integer(0) # Expected

c(1:10, 100L, 100L) |> unique_outliers()
#> [1] 100 # Expected

c(1:10, 100L) |> unique_outliers(n_iqr = 1000)
#> integer(0) # Expected
}
\seealso{
Other utility functions: 
\code{\link{closest_geobr_year}()},
\code{\link{fix_postal_code}()},
\code{\link{remove_unique_outliers}()},
\code{\link{test_date_line}()},
\code{\link{test_geobr_connection}()}
}
\concept{utility functions}
