% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_download.R
\name{worldclim_download}
\alias{worldclim_download}
\title{Download WorldClim data}
\usage{
worldclim_download(
  series,
  resolution = NULL,
  variable = NULL,
  model = NULL,
  ssp = NULL,
  year = NULL,
  dir = here::here("data")
)
}
\arguments{
\item{series}{A \code{\link[base:character]{character}} vector with the name of the
WorldClim data series (default: \code{"hcd"}).
The following options are available:
\itemize{
\item \code{"hcd"} = Historical Climate Data
\item \code{"hmwd"} = Historical Monthly Weather Data
\item \code{"fcd"} = Future Climate Data
}}

\item{resolution}{(optional) A \code{\link[base:character]{character}} vector with
the resolution of the WorldClim data series (default: \code{NULL}).
The following options are available:
\itemize{
\item \code{"10m"} = 10 Minutes (~340 km2 at the Equator)
\item \code{"5m"} = 5 Minutes (~85 km2 at the Equator)
\item \code{"2.5m"} = 2.5 Minutes (~21 km2 at the Equator)
\item \code{"30s"} = 30 Seconds (~1 km2  at the Equator) (not available for the
\code{"hmwd"} series)
}}

\item{variable}{(optional) A \code{\link[base:character]{character}} vector with the
variable of the WorldClim data series (default: \code{NULL}).

The following options are available for the \emph{Historical Climate Data}
series:
\itemize{
\item \code{"tmin"} = Average Minimum Temperature (°C)
\item \code{"tavg"} = Average Temperature (°C)
\item \code{"tmax"} = Average Maximum Temperature (°C)
\item \code{"srad"} = Solar Radiation (kJ m^-2 day^-1)
\item \code{"prec"} = Total Precipitation (mm)
\item \code{"vapr"} = Water Vapor Pressure (kPa)
\item \code{"wind"} = Wind Speed (m s^-1)
\item \code{"bioc"} = Bioclimatic Variables
\item \code{"elev"} = Elevation
}

The following options are available for the \emph{Historical Monthly Weather
Data} series:
\itemize{
\item \code{"tmin"} = Average Minimum Temperature (°C)
\item \code{"tmax"} = Average Maximum Temperature (°C)
\item \code{"prec"} = Total Precipitation (mm)
}

The following options are available for the \emph{Future Climate Data} series:
\itemize{
\item \code{"tmin"} = Average Minimum Temperature (°C)
\item \code{"tmax"} = Average Maximum Temperature (°C)
\item \code{"prec"} = Total Precipitation (mm)
\item \code{"bioc"} = Bioclimatic Variables
}}

\item{model}{(optional) A \code{\link[base:character]{character}} vector with the
global climate model of the WorldClim data series (default: \code{NULL}).

The following options are available for the \emph{Future Climate Data} series:
\itemize{
\item \code{"ACCESS-CM2"} = Australian Community Climate and Earth-System Simulator,
Australia
\item \code{"BCC-CSM2-MR"} = Beijing Climate Center Climate System Model, China
\item \code{"CMCC-ESM2"} = Canadian Earth System Model, Italy
\item \code{"EC-Earth3-Veg"} = European Consortium Earth System Model, Europe
\item `"FIO-ESM-2-0"`` = First Institute of Oceanography Earth System Model,
China
\item \code{"GFDL-ESM4"} = Geophysical Fluid Dynamics Laboratory Earth System Model,
USA
\item \code{"GISS-E2-1-G"} = NASA Goddard Institute for Space Studies Model, USA
\item \code{"HadGEM3-GC31-LL"} = Hadley Global Environment Model, UK
\item \code{"INM-CM5-0"} = Institute of Numerical Mathematics Climate Model, Russia
\item \code{"IPSL-CM6A-LR"} = Institut Pierre-Simon Laplace Climate Model, France
\item \code{"MIROC6"} = Model for Interdisciplinary Research on Climate, Japan
\item \code{"MPI-ESM1-2-HR"} = Max Planck Institute for Meteorology Earth System
Model, Germany
\item \code{"MRI-ESM2-0"} = Meteorological Research Institute Earth System Model,
Japan
\item \code{"UKESM1-0-LL"} = UK Earth System Model, UK
}}

\item{ssp}{(optional) A \code{\link[base:character]{character}} vector with the
shared socioeconomic pathway (SSP) of the WorldClim data series
(default: \code{NULL}).

The following options are available for the \emph{Future Climate Data} series:
\itemize{
\item \code{"ssp126"} = Sustainability (Taking the Green Road)
\item \code{"ssp245"} = Middle of the Road
\item \code{"ssp370"} = Regional Rivalry (A Rocky Road)
\item \code{"ssp585"} = Fossil-fueled Development (Taking the Highway)
}}

\item{year}{(optional) A \code{\link[base:character]{character}} vector with the
year group of the WorldClim data series (default: \code{NULL}). This is only
available for the \emph{Historical Monthly Weather Data} and \emph{Future Climate
Data} series:
\itemize{
\item \emph{Historical Monthly Weather Data}: \code{"1960-1959"}, \code{"1970-1979"},
\code{"1980-1989"}, \code{"1990-1999"},
\code{"2000-2009"}, \code{"2010-2019"}, \code{"2020-2024"}
\item \emph{Future Climate Data}: \code{"2021-2040"}, \code{"2041-2060"}, \code{"2061-2080"},
\code{"2081-2100"}
}}

\item{dir}{A string specifying the directory where to save the downloaded
files (default: \code{here::here("data")}).}
}
\value{
An invisible \code{\link[base:character]{character}} vector with the file
path(s) of the downloaded data.
}
\description{
\code{worldclim_download()} downloads and unzips data from the
\href{https://worldclim.org/}{WorldClim} website.

See \code{\link[geodata:worldclim]{worldclim_global()}},
\code{\link[geodata:worldclim]{worldclim_country()}}, and
\code{\link[geodata:worldclim]{worldclim_tile()}} from the
\href{https://cran.r-project.org/package=geodata}{geodata} package for
alternative ways to download WorldClim data.

\strong{Note:} This function requires an active internet connection and the
\href{https://CRAN.R-project.org/package=curl}{\code{curl}},
\href{https://CRAN.R-project.org/package=fs}{\code{fs}},
\href{https://CRAN.R-project.org/package=httr}{\code{httr}},
\href{https://CRAN.R-project.org/package=rvest}{\code{rvest}}, and
\href{https://CRAN.R-project.org/package=zip}{\code{zip}} packages to be
installed.
}
\examples{
\dontrun{
  if (FALSE) {
    worldclim_download(
      series = "hcd",
      resolution = "10m",
      variable = "prec"
    )
  }
}

\dontrun{
  if (FALSE) {
    worldclim_download(
      series = "fcd",
      resolution = "10m",
      variable = "tmin",
      model = "ACCESS-CM2",
      ssp = "ssp245",
      year = "2041-2060"
    )
  }
}
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_extract_variable}()},
\code{\link{worldclim_file}()},
\code{\link{worldclim_random}()},
\code{\link{worldclim_to_ascii}()},
\code{\link{worldclim_url}()}
}
\concept{WorldClim functions}
