% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_random.R
\name{worldclim_random}
\alias{worldclim_random}
\title{Return a random WorldClim data selection}
\usage{
worldclim_random(series = "hcd")
}
\arguments{
\item{series}{A string indicating the \href{https://worldclim.org/}{WorldClim}
data series to use. The following options are available:
\itemize{
\item \code{'hcd'} (Historical Climate Data).
\item \code{'hmwd'} (Historical Monthly Weather Data).
\item \code{'fcd'} (Future Climate Data).
}}
}
\value{
A \code{\link[base:list]{list}} with a random selection of parameters for
downloading WorldClim data. The list contains the following elements:
\itemize{
\item \code{series}: The selected data series.
\item \code{resolution}: The selected spatial resolution.
\item \code{variable}: The selected climate variable.
\item \code{bioclimatic_variable}: The selected bioclimatic variable (only for
Historical Climate Data and Future Climate Data when \code{variable} is
set to \code{"bioc"}).
\item \code{model}: The selected climate model (only for Future Climate Data).
\item \code{ssp}: The selected Shared Socioeconomic Pathway (only for Future
Climate Data).
\item \code{year}: The selected year or year range.
\item \code{month}: The selected month.
}
}
\description{
\code{worldclim_random()} returns a random selection of parameters for
downloading \href{https://worldclim.org/}{WorldClim} data.
}
\examples{
worldclim_random("hcd")

worldclim_random("hmwd")

worldclim_random("fcd")
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_download}()},
\code{\link{worldclim_extract_variable}()},
\code{\link{worldclim_file}()},
\code{\link{worldclim_to_ascii}()},
\code{\link{worldclim_url}()}
}
\concept{WorldClim functions}
