% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_fu.R
\name{get_brazil_fu}
\alias{get_brazil_fu}
\title{Get Brazilian federative unit abbreviations}
\usage{
get_brazil_fu(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names or numeric codes of Brazilian regions or federal units.
Municipality codes are also supported. If \code{NULL},
returns a \code{\link[base:character]{character}} vector with all Brazilian
federal unit abbreviations (default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the abbreviations of
Brazilian federal units.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_fu()} returns a vector with the abbreviations of
Brazilian federal units.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
get_brazil_fu()

get_brazil_fu("sp")
#> [1] "SP" # Expected

get_brazil_fu("sao paulo")
#> [1] "SP" # Expected

get_brazil_fu(35)
#> [1] "SP" # Expected

get_brazil_fu(3550308) # Municipality of São Paulo
#> [1] "SP" # Expected

get_brazil_fu(35503081) # >7 digits
#> [1] NA # Expected

get_brazil_fu(39027001) # Non-existent state code
#> [1] NA # Expected

get_brazil_fu("southeast")
#> [1] "ES" "MG" "RJ" "SP" # Expected

get_brazil_fu(3)
#> [1] "ES" "MG" "RJ" "SP" # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
