% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_code.R
\name{get_brazil_state_code}
\alias{get_brazil_state_code}
\title{Get Brazilian state codes}
\usage{
get_brazil_state_code(x = NULL)
}
\arguments{
\item{x}{(Optional) A \code{\link[base:character]{character}} vector with the names
of Brazilian states or federal units. If \code{NULL}, returns all state codes
(Default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the IBGE codes of
Brazilian states.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_brazil_state_code()} returns a vector with the Brazilian Institute
of Geography and Statistics (\href{https://www.ibge.gov.br/}{IBGE}) codes for
Brazilian states.
}
\details{
The data from this function is based on data from the Brazilian Institute
of Geography and Statistics (IBGE, n.d.).
}
\examples{
get_brazil_state_code()

get_brazil_state_code("acre")
#> [1] 12 # Expected

get_brazil_state_code("ac")
#> [1] 12 # Expected
}
\references{
Instituto Brasileiro de Geografia e Estatística. (n.d.). \emph{Território}
[Territory] [Dataset]. SIDRA. \url{https://sidra.ibge.gov.br/territorio}
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
