% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_state_latitude.R
\name{get_brazil_state_latitude}
\alias{get_brazil_state_latitude}
\title{Get Brazilian state latitude}
\usage{
get_brazil_state_latitude(x = NULL)
}
\arguments{
\item{x}{(optional) An \code{\link[base:is.recursive]{atomic}} vector containing the
names of Brazilian states or federal units. Municipality and state codes
are also supported. If \code{NULL}, returns a vector with all state latitudes
(default: \code{NULL}).}
}
\value{
A \code{\link[base:character]{character}} vector with the latitude of
Brazilian state capitals.
}
\description{
\code{get_brazil_state_latitude()} returns a vector with the latitude of
Brazilian state capitals.
}
\details{
The data from this function is based on Google's Geocoding API gathered via
the \code{\link[tidygeocoder:tidygeocoder-package]{tidygeocoder}} R package.
}
\examples{
get_brazil_state_latitude()

get_brazil_state_latitude("sp")
#> [1] -23.55065 # Expected

get_brazil_state_latitude("sao paulo")
#> [1] -23.55065 # Expected

get_brazil_state_latitude(35) # State of São Paulo
#> [1] -23.55065 # Expected

get_brazil_state_latitude(3550308) # Municipality of São Paulo
#> [1] -23.55065 # Expected

get_brazil_state_latitude(35503081) # >7 digits
#> [1] NA # Expected

get_brazil_state_latitude(3912345) # Non-existent state code
#> [1] NA # Expected
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_longitude}()},
\code{\link{render_brazil_address}()}
}
\concept{Brazil functions}
