% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbis-package.R
\docType{package}
\name{orbis-package}
\alias{orbis}
\alias{orbis-package}
\title{orbis: Spatial Data Analysis Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A suite of tools for simplifying spatial data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/orbis/}
  \item \url{https://github.com/danielvartan/orbis/}
  \item Report bugs at \url{https://github.com/danielvartan/orbis/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danielvartan@proton.me} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
