% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_brazil_address.R
\name{render_brazil_address}
\alias{render_brazil_address}
\title{Render Brazilian addresses}
\usage{
render_brazil_address(
  street = NA_character_,
  complement = NA_character_,
  neighborhood = NA_character_,
  municipality = NA_character_,
  state = NA_character_,
  postal_code = NA_character_
)
}
\arguments{
\item{street}{(optional) A \code{\link[base:character]{character}} vector with the
street names (default: \code{NA_character}).}

\item{complement}{(optional) A \code{\link[base:character]{character}} vector with
the complement of the address (default: \code{NA_character}).}

\item{neighborhood}{(optional) A \code{\link[base:character]{character}} vector with
the neighborhood names (default: \code{NA_character}).}

\item{municipality}{(optional) A \code{\link[base:character]{character}} vector
with the name of the municipalities (default: \code{NA_character}).}

\item{state}{(optional) A \code{\link[base:character]{character}} vector with the
name of the states (default: \code{NA_character}).}

\item{postal_code}{(optional) A \code{\link[base:character]{character}} vector with
the postal codes (default: \code{NA_character}).}
}
\value{
A \code{\link[base:character]{character}} vector with the formatted address.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{render_brazil_address()}
}
\examples{
render_brazil_address(
  street = c("Viaduto do Chá, 15", "Alameda Ribeiro da Silva, 919"),
  complement = c("", "Ap. 502"),
  neighborhood = c("Centro", "Campos Elíseos"),
  municipality = c("São Paulo", "São Paulo"),
  state = c("SP", "SP"),
  postal_code = c("01002-020", "01217-010")
)
}
\seealso{
Other Brazil functions: 
\code{\link{get_brazil_fu}()},
\code{\link{get_brazil_municipality}()},
\code{\link{get_brazil_municipality_code}()},
\code{\link{get_brazil_region}()},
\code{\link{get_brazil_region_code}()},
\code{\link{get_brazil_state}()},
\code{\link{get_brazil_state_by_utc}()},
\code{\link{get_brazil_state_capital}()},
\code{\link{get_brazil_state_code}()},
\code{\link{get_brazil_state_latitude}()},
\code{\link{get_brazil_state_longitude}()}
}
\concept{Brazil functions}
