% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_extract.R
\name{worldclim_extract_variable}
\alias{worldclim_extract_variable}
\alias{worldclim_extract_resolution}
\alias{worldclim_extract_month}
\alias{worldclim_extract_year}
\alias{worldclim_extract_year_month}
\alias{worldclim_extract_year_group}
\alias{worldclim_extract_gcm}
\alias{worldclim_extract_ssp}
\title{Extract components from WorldClim filenames}
\usage{
worldclim_extract_variable(file)

worldclim_extract_resolution(file)

worldclim_extract_month(file)

worldclim_extract_year(file)

worldclim_extract_year_month(file)

worldclim_extract_year_group(file)

worldclim_extract_gcm(file)

worldclim_extract_ssp(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector with WorldClim
filenames.}
}
\value{
A \code{\link[base:character]{character}} vector with the extracted
component.
}
\description{
\verb{worldclim_extract_*()} extract components present in WorldClim filenames.
}
\examples{
files <- c(
  "wc2.1_10m_tavg_1970-2000-06.asc",
  "wc2.1_cruts4.09_30s_tmin_1962-10.asc",
  "wc2.1_5m_tmax_GISS-E2-1-G_ssp370_2081-2100-05.asc"
)

files |> worldclim_extract_variable()
#> [1] "tavg" "tmin" "tmax" # Expected

files |> worldclim_extract_resolution()
#> [1] "10m" "30s" "5m" # Expected

files |> worldclim_extract_month()
#> [1] "06" "10" "05" # Expected

files |> worldclim_extract_year()
#> [1] "1970-2000" "1962"      "2081-2100" # Expected

files |> worldclim_extract_year_month()
#> [1] "1970-2000-06" "1962-10"      "2081-2100-05" # Expected

files |> worldclim_extract_year_group()
#> [1] "1970-2000" NA          "2081-2100" # Expected

files |> worldclim_extract_gcm()
#> [1] NA            NA            "GISS-E2-1-G" # Expected

files |> worldclim_extract_ssp()
#> [1] NA       NA       "ssp370" # Expected
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_random}()},
\code{\link{worldclim_to_ascii}()},
\code{\link{worldclim_url}()}
}
\concept{WorldClim functions}
