% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_to_ascii.R
\name{worldclim_to_ascii}
\alias{worldclim_to_ascii}
\title{Convert WorldClim GeoTIFF files to Esri ASCII Grid}
\usage{
worldclim_to_ascii(
  file,
  shape = NULL,
  box = NULL,
  shift_longitude = TRUE,
  dx = -45,
  remove_extreme_outliers = TRUE,
  overwrite = TRUE,
  na_flag = -99,
  dir = dirname(file[1])
)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector of file paths to the
\href{https://worldclim.org/}{WorldClim}
\href{https://en.wikipedia.org/wiki/GeoTIFF}{GeoTIFF} files to be converted.
The files must have a \code{.tif} extension.}

\item{shape}{(optional) An \code{\link[sf:sf]{sf}} or
\code{\link[terra:vect]{SpatVector}} class object representing the polygon to crop
the raster data (default: \code{NULL}).}

\item{box}{(optional) A \code{\link[base:numeric]{numeric}} vector of length 4
specifying the bounding box for cropping the raster data in the format
\code{c(xmin, ymin, xmax, ymax)} (default: \code{NULL}).}

\item{shift_longitude}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to apply a date line fix to the raster data when
the shape parameter is provided and spans the
\href{https://en.wikipedia.org/wiki/International_Date_Line}{International Date Line}.
This is particularly useful when working with rasters and vectors that span
the date line (e.g. the Russian territory). See
\code{\link[=shift_and_rotate]{shift_and_rotate()}} to learn more (default: \code{TRUE}).}

\item{dx}{(optional) A \code{\link[base:numeric]{numeric}} value specifying the
horizontal distance in degrees to shift the raster data. This is only
relevant if \code{shift_longitude} is set to \code{TRUE} (default: \code{-45}).}

\item{remove_extreme_outliers}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to transform to \code{NA} values 10 times the interquartile
range (\href{https://en.wikipedia.org/wiki/Interquartile_range}{IQR}) below the
first quartile or above the third quartile of the data values without
duplications. This is useful to remove abnormal values in the raster data
(default: \code{FALSE}).}

\item{overwrite}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to overwrite existing files in the output directory
(default: \code{TRUE}).}

\item{na_flag}{(optional) An \code{\link[base:integer]{integer}} value specifying
the \code{NODATA_VALUE} for the output ASCII files. See the \emph{Details} section
to learn more (default: \code{-99}).}

\item{dir}{(optional) A \code{\link[base:character]{character}} vector specifying
the output directory for the converted
\href{https://en.wikipedia.org/wiki/Esri_grid}{Esri ASCII Grid}
files. Defaults to the directory of the first file in the \code{file} parameter
(default: \code{dirname(file[1])}).}
}
\value{
An invisible \code{\link[base:character]{character}} vector containing the
file paths of the converted ASCII files.
}
\description{
\code{worldclim_to_ascii()} facilitates the conversion of one or more
\href{https://worldclim.org/}{WorldClim}
\href{https://en.wikipedia.org/wiki/GeoTIFF}{GeoTIFF} files to the
\href{https://en.wikipedia.org/wiki/Esri_grid}{Esri ASCII Grid} raster format.
}
\details{
\subsection{\code{na_flag} parameter}{

According to the \href{https://en.wikipedia.org/wiki/Esri_grid}{Esri ASCII}
raster format documentation, the default value for \code{NODATA_VALUE}
(the \code{NA} flag) is \code{-9999}. However, using four digits of precision
significantly inflates file size. For
\href{https://worldclim.org/}{WorldClim} data, two significant digits (\code{-99}) are
sufficient, since the only variables with negative values are temperatures,
and the lowest temperature ever recorded on Earth is above that threshold.
}
}
\examples{
# Set the Environment -----

library(fs)
library(httr2)
library(magrittr)
library(readr)
library(rvest)
library(zip)

# Download a WorldClim Dataset -----

\dontrun{
  if (is_online()) {
    tif_file <-
      worldclim_download(
        series = "hcd",
        resolution = "10m",
        variable = "prec",
        dir = tempdir()
      ) |>
        magrittr::extract(1)
  }
}

# Transform Data to Esri ASCII -----

\dontrun{
  if (is_online()) {
    asc_file <- tif_file |> worldclim_to_ascii()
  }
}

# Check the Output -----

\dontrun{
  if (is_online()) {
    asc_file |> read_lines(n_max = 11)
  }
}
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_download}()},
\code{\link{worldclim_extract_variable}()},
\code{\link{worldclim_file}()},
\code{\link{worldclim_random}()},
\code{\link{worldclim_url}()}
}
\concept{WorldClim functions}
