% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_url.R
\name{worldclim_url}
\alias{worldclim_url}
\title{Get paths to WorldClim data}
\usage{
worldclim_url(series, resolution = "all")
}
\arguments{
\item{series}{A \code{\link[base:character]{character}} vector with the name of the
\href{https://worldclim.org/}{WorldClim} data series (default: \code{"hcd"}).
The following options are available:
\itemize{
\item \code{'hcd'} (Historical Climate Data).
\item \code{'hmwd'} (Historical Monthly Weather Data).
\item \code{'fcd'} (Future Climate Data).
}}

\item{resolution}{(optional) A \code{\link[base:character]{character}} vector with
the resolution of the \href{https://worldclim.org/}{WorldClim} data series
(default: \code{"all"}). The following options are available:
\itemize{
\item \code{"all"} returns all available resolutions.
\item \code{"10m"} returns the 10m resolution.
\item \code{"5m"} returns the 5m resolution.
\item \code{"2.5m"} returns the 2.5m resolution.
\item \code{"30s"} returns the 30s resolution.
}}
}
\value{
A named \code{\link[base:character]{character}} vector with the URL(s) of
the \href{https://worldclim.org/}{WorldClim} data series.
}
\description{
\code{worldclim_url()} returns the URL(s) of a
\href{https://worldclim.org/}{WorldClim} data series.
}
\examples{
worldclim_url("hcd")

worldclim_url("hmwd")

worldclim_url("fcd")

worldclim_url("fcd", "10m")

worldclim_url(c("hcd", "hmwd", "fcd"))

worldclim_url(c("hcd", "hmwd", "fcd"), "5m")
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_extract_variable}()},
\code{\link{worldclim_random}()},
\code{\link{worldclim_to_ascii}()}
}
\concept{WorldClim functions}
