% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_url.R
\name{worldclim_url}
\alias{worldclim_url}
\title{Get paths to WorldClim data}
\usage{
worldclim_url(series, resolution = "all")
}
\arguments{
\item{series}{A \code{\link[base:character]{character}} vector with the name of the
WorldClim data series. The following options are available:
\itemize{
\item \code{"hcd"} or \code{"historical-climate-data"} or \code{"historical climate data"}.
\item \code{"hmwd"} or \code{"historical-monthly-weather-data"} or
\code{"historical monthly weather data"}.
\item \code{"fcd"} or \code{"future-climate-data"} or \code{"future climate data"}.
}}

\item{resolution}{(optional) A \code{\link[base:character]{character}} vector with
the resolution of the WorldClim data series. The following options are
available:
\itemize{
\item \code{"all"} (default) returns all available resolutions.
\item \code{"10m"} returns the 10m resolution.
\item \code{"5m"} returns the 5m resolution.
\item \code{"2.5m"} returns the 2.5m resolution.
\item \code{"30s"} returns the 30s resolution.
}}
}
\value{
A named \code{\link[base:character]{character}} vector with the URL(s) or
OSF ID(s) of the WorldClim data series.
}
\description{
\code{worldclim_url()} returns the URL(s) of a WorldClim data series.
}
\examples{
worldclim_url("hcd")

worldclim_url("historical climate data")

worldclim_url("hmwd")

worldclim_url("historical monthly weather data")

worldclim_url("fcd")

worldclim_url("future climate data")

worldclim_url("future climate data", "10m")

worldclim_url(c("hcd", "hmwd", "fcd"))

worldclim_url(c("hcd", "hmwd", "fcd"), "5m")
}
\seealso{
Other WorldClim functions: 
\code{\link{worldclim_extract_variable}()},
\code{\link{worldclim_random}()},
\code{\link{worldclim_to_ascii}()}
}
\concept{WorldClim functions}
